Acabar pokemon.php y pokemon_edit.php

+++ php_controllers/pokemonController.php
@@ -1,41 +1,44 @@
 <?php
 include '../php_librarys/pokedex.php';
-
-if(isset($_SESSION['pokedex'])) {
+if(!isset($_SESSION)){
+    session_start();
+   }
+if (isset($_SESSION['pokedex'])) {
     $pokedex = $_SESSION['pokedex'];
 } else {
     $pokedex = [];
 }
 //Ruta temporal de la imagen 
 
-
-$rutaTemporal = $_FILES['imagen']['tmp_name'];
-if(isset($_FILES['imagen'])){
-   move_uploaded_file($rutaTemporal,$_FILES['imagen']['name']);
-}
-//Creamos el pokemon
-$imagen = $_FILES['imagen']['name'];
-$pokemon = crearPokemon($_POST['txtNumero'],$_POST['txtNombre'],$_POST['region'],$_POST['tipoPokemon'],$_POST['altura'],$_POST['peso'],$_POST['rEvolucion'],$imagen);
-
-$pokedex = addPokemon($pokedex,$pokemon);
-if($_SESSION['addPokemon' ] == 'Pokemon añadido correctamente') {
-    move_uploaded_file($rutaTemporal,"../media/");
-    header("Location: ../php_views/pokemon_list.php");
-    exit();
-} else {
-    $_SESSION['pokemon'] = $pokemon;
-    header("Location: ../php_views/pokemon.php");
-    exit();
-}
-$pokedex = deletePokemon($pokedex,$_POST['txtNumero']);
-if($_SESSION['deletePokemon'] == 'El pokemon seleccionado se ha borrado de la pokedex') {
-    $indexOfPokemon = array_search($_POST['txtNumero'], array_column($pokedex, 'Numero'));
-    unlink("../media/".$pokemon[$indexOfPokemon]['Imatge']);
-    $_SESSION['pokedex'] = $pokedex;
-} else {
-    $_SESSION['deletePokemon'] == "No se ha borrado el pokemon.Problemas al borrar la imagen";
-}
-
-
-
-
+if ($_POST['add']) {
+    $rutaTemporal = $_FILES['imagen']['tmp_name'];
+    if (isset($_FILES['imagen'])) {
+        move_uploaded_file($rutaTemporal, $_FILES['imagen']['name']);
+    }
+    //Creamos el pokemon
+    $imagen = $_FILES['imagen']['name'];
+    $pokemon = crearPokemon($_POST['txtNumero'], $_POST['txtNombre'], $_POST['region'], $_POST['tipoPokemon'], $_POST['altura'], $_POST['peso'], $_POST['rEvolucion'], $imagen);
+    $pokedex = addPokemon($pokedex, $pokemon);
+    if ($_SESSION['addPokemon'] == 'Pokemon añadido correctamente') {
+        move_uploaded_file($rutaTemporal, "../media/");
+        header('Location:' . '../php_views/pokemon_list.php');
+        exit();
+    } else {
+        $_SESSION['pokemon'] = $pokemon;
+        header('Location:' . '../php_views/pokemon.php');
+        exit();
+    }
+} /*elseif ($_POST['edit']) {
+    $pokemon = modifyPokemon($pokedex,$_POST['txtNumero'],$_POST['txtNumero'], $_POST['txtNombre'], $_POST['region'], $_POST['tipoPokemon'], $_POST['altura'], $_POST['peso'], $_POST['rEvolucion'], $imagen);
+    $_SESSION['pokemonModified'] = $pokemon;
+    header('Location:' . '../php_views/pokemon_edit.php');
+} elseif ($_POST['delete']) {
+    $pokedex = deletePokemon($pokedex, $_POST['txtNumero']);
+    if ($_SESSION['deletePokemon'] == 'El pokemon seleccionado se ha borrado de la pokedex') {
+        $indexOfPokemon = array_search($_POST['txtNumero'], array_column($pokedex, 'Numero'));
+        unlink("../media/" . $pokemon[$indexOfPokemon]['Imatge']);
+        $_SESSION['pokedex'] = $pokedex;
+    } else {
+        $_SESSION['deletePokemon'] == "No se ha borrado el pokemon.Problemas al borrar la imagen";
+    }
+}*/
